requirements.txt
fastapi
uvicorn
python-dotenv
aiomysql
tortoise-orm
aerich
python-jose[cryptography]
bcrypt
passlib
pydantic[email]
python-multipart